fastapi/requirements.txt
# alembic==1.15.1
# amqp==5.3.1
# aniso8601==10.0.0
# ansible==10.7.0
# ansible-base==2.10.8
# ansible-compat==25.1.4
# ansible-core==2.17.9
# ansible-lint==25.1.3
# appdirs==1.4.4
# argcomplete==1.8.1
# asgiref==3.8.1
# astroid==3.3.8
# async-timeout==5.0.1
# attrs==25.1.0
# Automat==20.2.0
# Babel==2.8.0
# bcrypt==3.2.0
# billiard==4.2.1
# black==25.1.0
# blinker==1.9.0
# bracex==2.5.post1
# Brotli==1.1.0
# cachelib==0.13.0
# celery==5.4.0
# certifi==2020.6.20
# cffi==1.17.1
# chardet==4.0.0
# charset-normalizer==3.4.1
# ciscoisesdk==2.3.0
# click==8.1.8
# click-didyoumean==0.3.1
# click-plugins==1.1.1
# click-repl==0.3.0
# colorama==0.4.4
# configobj==5.0.6
# constantly==15.1.0
# coverage==7.6.12
# crontab==1.0.1
# cryptography==44.0.2
# dbus-python==1.2.18
# deepdiff==8.3.0
# defusedxml==0.7.1
# dict2xml==1.7.6
# diff-match-patch==20241021
# dill==0.3.9
# distlib==0.3.9
# distro==1.7.0
# Django==5.1.7
# django-allauth==65.4.1
# django-celery==3.1.17
# django-cors-headers==4.7.0
# django-crispy-forms==2.3
# django-debug-toolbar==5.0.1
# django-environ==0.12.0
# django-extensions==3.2.3
# django-import-export==4.3.7
# django-redis==5.4.0
# django-taggit==6.1.0
# djangorestframework==3.15.2
# dnspython==2.1.0
# docker==7.1.0
# #docker-compose==1.28.5
# dockerpty==0.4.1
# docopt==0.6.2
# dpkt==1.9.8
# et_xmlfile==2.0.0
# exceptiongroup==1.2.2
# # f5-icontrol-rest==1.3.13
# fake-factory==9999.9.9
# fastjsonschema==2.21.1
# # filelock==3.17.0
# # Flask==3.1.0
# # Flask-Admin==1.6.1
# # Flask-Bcrypt==1.0.1
# # Flask-Caching==2.3.1
# # Flask-Celery-Helper==1.1.0
# # Flask-Compress==1.17
# # flask-cors==5.0.1
# # Flask-DebugToolbar==0.16.0
# # Flask-FileUpload==0.5.0
# # Flask-JWT-Extended==4.7.1
# # Flask-Login==0.6.3
# # Flask-Migrate==4.1.0
# # flask-peewee==3.0.6
# # flask-redis==0.4.0
# # Flask-RESTful==0.3.10
# # Flask-RQ2==18.3
# # Flask-Session==0.8.0
# # Flask-SQLAlchemy==3.1.1
# # Flask-Testing==0.8.1
# # Flask-Uploads==0.2.1
# # Flask-WTF==1.2.2
# freezegun==1.5.1
# future==1.0.0
# genie==25.1
# genie.libs.clean==25.1
# genie.libs.conf==25.1
# genie.libs.filetransferutils==25.1
# genie.libs.health==25.1
# genie.libs.ops==25.1
# genie.libs.parser
# genie.libs.sdk==25.1
# greenlet==3.1.1
# grpcio==1.70.0
# gunicorn==23.0.0
# httplib2==0.20.2
# hyperlink==21.0.0
# idna==3.3
# importlib-metadata==4.6.4
# importlib_resources==6.5.2
# incremental==21.3.0
# influxdb==5.3.2
# iniconfig==2.0.0
# isort==6.0.1
# itsdangerous==2.2.0
# jdcal==1.4.1
# jeepney==0.7.1
# Jinja2==3.1.6
# jmespath==0.10.0
# jsonpatch==1.32
# jsonpickle==4.0.2
# jsonpointer==2.0
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
# keyring==23.5.0
# kombu==5.4.2
# launchpadlib==1.10.16
# lazr.restfulclient==0.14.4
# lazr.uri==1.0.6
# lazy-object-proxy==1.10.0
# ldap3==2.9.1
# leveldb==0.201
# lockfile==0.12.2
# lxml==5.3.1
# Mako==1.3.9
# markdown-it-py==3.0.0
# MarkupPy==1.14
# MarkupSafe==3.0.2
# mccabe==0.7.0
# mdurl==0.1.2
# more-itertools==8.10.0
# msgpack==1.1.0
# msgspec==0.19.0
# mypy-extensions==1.0.0
# #ncclient==0.6.19
# netaddr==1.3.0
# netifaces==0.11.0
# netmiko==4.5.0
# ntc_templates==7.7.0
# ntlm-auth==1.4.0
# oauthlib==3.2.0
# odfpy==1.4.1
# openpyxl==3.1.5
# ordered-set==4.1.0
# orderly-set==5.3.0
# packaging==24.2
# paramiko==3.5.1
# pathspec==0.12.1
# peewee==3.17.9
# # ping3==4.0.8
# platformdirs==4.3.6
# pluggy==1.5.0
# ply==3.11
# prettytable==3.15.1
# prompt_toolkit==3.0.50
# protobuf==6.30.0
psycopg2-binary==2.9.10
# pyasn1==0.4.8
# pyasn1-modules==0.2.1
# pycparser==2.22
# pycryptodomex==3.11.0
# pyftpdlib==2.0.1
# Pygments==2.19.1
# PyHamcrest==2.0.2
# PyJWT==2.3.0
# pykerberos==1.1.14
# pylint==3.3.4
# pymsteams==0.2.5
# PyNaCl==1.5.0
# pyOpenSSL==21.0.0
# pyparsing==2.4.7
# pyrsistent==0.18.1
# pyserial==3.5
# pysmi==1.5.11
# pysnmp==6.1.4
# pysnmpcrypto==0.0.4
# pytest==8.3.5
# pytest-cov==4.1.0
# pytest-django==4.10.0
# python-dateutil==2.9.0.post0
# python-dotenv==0.19.2
# pywinrm==0.3.0
# redis==5.2.1
# referencing==0.36.2
# requests==2.32.3
# requests-kerberos==0.12.0
# requests-ntlm==1.1.0
# requests-toolbelt==1.0.0
# resolvelib==1.0.1
# rest.connector==25.1
# rich==13.9.4
# rpds-py==0.23.1
# rq==2.1.0
# rq-scheduler==0.14.0
# ruamel.yaml==0.18.10
# ruamel.yaml.clib==0.2.12
# scp==0.15.0
# SecretStorage==3.3.1
# service-identity==18.1.0
# simplejson==3.17.6
# six==1.16.0
# snmp-cmds==1.0
# SQLAlchemy==2.0.38
# sqlparse==0.5.3
# subprocess-tee==0.4.2
# tablib==3.8.0
# tenacity==9.0.0
# # textfsm==1.1.3
# # texttable==1.6.4
# tftpy==0.8.0
# toml==0.10.2
# tomli==2.2.1
# tomlkit==0.13.2
# tqdm==4.67.1
# Twisted==22.1.0
# typed-ast==1.5.5
# typing_extensions==4.12.2
# tzdata==2025.1
# unicon==25.1
# unicon.plugins==25.1
# urllib3==1.26.5
# vine==5.1.0
# virtualenv==20.29.3
# wadllib==1.3.6
# wcmatch==10.0
# wcwidth==0.2.13
# websocket-client==1.2.3
# Werkzeug==3.1.3
# wrapt==1.17.2
# wtf-peewee==3.0.6
# WTForms==3.2.1
# xlrd==2.0.1
# xlwt==1.3.0
# xmltodict==0.12.0
# yamllint==1.35.1
# yang.connector==25.1
# zipp==1.0.0
# zope.interface==5.4.0
# zstandard==0.23.0

# FastAPI 기본 패키지
setuptools<71
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.23.0
python-dotenv==1.0.0
pyats==25.8
genie==25.1
textfsm==1.1.3
netmiko==4.3.0
pytz==2024.1
requests==2.32.3
fpdf2==2.8.3
python-dateutil==2.9.0.post0